% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipesGadget.R
\name{recipesGadget}
\alias{recipesGadget}
\title{Recipes Gadget}
\usage{
recipesGadget(.df, bake_list = NULL)
}
\arguments{
\item{.df}{data frame, data table or tibble on which the preprocessing steps will be trained.}

\item{bake_list}{A named list where each element is a data.frame, data.table or tibble on which the selected preprocessing steps can be applied.}
}
\value{
Returns a list containing the trained recipe as well as all the data to which the recipe has been applied within the gadget.
}
\description{
Gadget to use the data preprocessing 'recipes' package interactively.
}
\examples{
\dontrun{
recipesGadget(.df = mtcars, bake_list = list('mtcars1' = mtcars1))
}
}
\author{
Alberto Almuiña
}
